refactor(qr-scanner): split private helpers out of qr_scanner_screen (Refs #563)
- qr_scanner_screen.dart: 391 -> 270 LOC.
- lib/features/sync/presentation/widgets/qr_scanner_helpers.dart (new):
  QrPermissionDenied, QrScanTimeoutPrompt, QrScanFrameOverlay, QrFramePainter
  (renamed from private versions to allow cross-file import).
- QrScannerScreen + QrScannerTorchButton stay in the screen file (test surface unchanged).
